refactor: move grid watcher system to network internals
