+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/grid/WirelessGrid.java
@@ -4,9 +4,9 @@ import com.refinedmods.refinedstorage.api.autocrafting.preview.Preview;
 import com.refinedmods.refinedstorage.api.autocrafting.task.TaskId;
 import com.refinedmods.refinedstorage.api.grid.operations.GridOperations;
 import com.refinedmods.refinedstorage.api.grid.operations.NoopGridOperations;
-import com.refinedmods.refinedstorage.api.grid.watcher.GridWatcher;
-import com.refinedmods.refinedstorage.api.grid.watcher.GridWatcherManager;
-import com.refinedmods.refinedstorage.api.grid.watcher.GridWatcherManagerImpl;
+import com.refinedmods.refinedstorage.api.network.node.grid.GridWatcher;
+import com.refinedmods.refinedstorage.api.network.impl.node.grid.GridWatcherManager;
+import com.refinedmods.refinedstorage.api.network.impl.node.grid.GridWatcherManagerImpl;
 import com.refinedmods.refinedstorage.api.network.autocrafting.AutocraftingNetworkComponent;
 import com.refinedmods.refinedstorage.api.network.energy.EnergyNetworkComponent;
 import com.refinedmods.refinedstorage.api.network.storage.StorageNetworkComponent;
