+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/grid/WirelessGridOperations.java
@@ -3,7 +3,7 @@ package com.refinedmods.refinedstorage.common.grid;
 import com.refinedmods.refinedstorage.api.grid.operations.GridExtractMode;
 import com.refinedmods.refinedstorage.api.grid.operations.GridInsertMode;
 import com.refinedmods.refinedstorage.api.grid.operations.GridOperations;
-import com.refinedmods.refinedstorage.api.grid.watcher.GridWatcherManager;
+import com.refinedmods.refinedstorage.api.network.impl.node.grid.GridWatcherManager;
 import com.refinedmods.refinedstorage.api.resource.ResourceKey;
 import com.refinedmods.refinedstorage.api.storage.ExtractableStorage;
 import com.refinedmods.refinedstorage.api.storage.InsertableStorage;
