+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/storage/portablegrid/PortableGrid.java
@@ -5,9 +5,9 @@ import com.refinedmods.refinedstorage.api.autocrafting.task.TaskId;
 import com.refinedmods.refinedstorage.api.core.Action;
 import com.refinedmods.refinedstorage.api.grid.operations.GridOperations;
 import com.refinedmods.refinedstorage.api.grid.operations.NoopGridOperations;
-import com.refinedmods.refinedstorage.api.grid.watcher.GridWatcher;
-import com.refinedmods.refinedstorage.api.grid.watcher.GridWatcherManager;
-import com.refinedmods.refinedstorage.api.grid.watcher.GridWatcherManagerImpl;
+import com.refinedmods.refinedstorage.api.network.node.grid.GridWatcher;
+import com.refinedmods.refinedstorage.api.network.impl.node.grid.GridWatcherManager;
+import com.refinedmods.refinedstorage.api.network.impl.node.grid.GridWatcherManagerImpl;
 import com.refinedmods.refinedstorage.api.network.energy.EnergyStorage;
 import com.refinedmods.refinedstorage.api.resource.ResourceKey;
 import com.refinedmods.refinedstorage.api.storage.Actor;
