+++ refinedstorage-network-api/src/main/java/com/refinedmods/refinedstorage/api/network/node/grid/package-info.java
@@ -1,6 +1,6 @@
 @ParametersAreNonnullByDefault
 @FieldsAndMethodsAreNonnullByDefault
-package com.refinedmods.refinedstorage.api.grid.watcher;
+package com.refinedmods.refinedstorage.api.network.node.grid;
 
 import com.refinedmods.refinedstorage.api.core.FieldsAndMethodsAreNonnullByDefault;
