+++ refinedstorage-network/src/main/java/com/refinedmods/refinedstorage/api/network/impl/node/grid/GridWatcherRootStorageListener.java
@@ -1,5 +1,6 @@
-package com.refinedmods.refinedstorage.api.grid.watcher;
+package com.refinedmods.refinedstorage.api.network.impl.node.grid;
 
+import com.refinedmods.refinedstorage.api.network.node.grid.GridWatcher;
 import com.refinedmods.refinedstorage.api.resource.list.MutableResourceList;
 import com.refinedmods.refinedstorage.api.storage.Actor;
 import com.refinedmods.refinedstorage.api.storage.root.RootStorage;
